requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
DateTime==4.3
Flask==2.0.2
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy~=1.18.5
pytz==2021.3
requests==2.27.1
scikit-learn~=0.23.1
scipy~=1.5.0
sklearn==0.0
threadpoolctl~=2.1.0
urllib3==1.26.8
Werkzeug==2.0.2
zope.interface==5.4.0

wtforms~=3.0.1
python-dotenv~=0.19.2
pip~=20.1.1
toml~=0.10.1
zlib~=1.2.11
wheel~=0.34.2
openssl~=1.1.1g
cryptography~=2.9.2
py~=1.9.0
lxml~=4.5.2
tornado~=6.0.4
typing_extensions~=3.7.4.2
keyring~=21.2.1
pytest~=5.4.3
zope~=1.0
cython~=0.29.21
setuptools~=49.2.0
nose~=1.3.7
cffi~=1.14.0
numba~=0.50.1
redis~=4.1.1
deprecated~=1.2.13
path~=13.1.0
importlib_metadata~=1.7.0
packaging~=20.4
matplotlib~=3.2.2
mpmath~=1.1.0
blas~=1.0
sympy~=1.6.1
pkginfo~=1.5.0.1
six~=1.15.0
ipython~=7.16.1
distributed~=2.20.0
dask~=2.20.0
psutil~=5.7.0
cloudpickle~=1.5.0
backports~=1.0
pillow~=7.2.0
greenlet~=0.4.16
gevent~=20.6.2
chardet~=3.0.4
babel~=2.8.0
pyparsing~=2.4.7
ply~=3.11
pycparser~=2.20
wrapt~=1.11.2
future~=0.18.2
docutils~=0.16
sqlalchemy~=1.3.18
attrs~=19.3.0
mock~=4.0.2
pyodbc~=4.0.30
bcrypt~=3.1.7